+++ tools/libprep/run.js
@@ -10,6 +10,7 @@
 var buildify = require('buildify'),
 path = require('path'),
 fs = require('fs-extra'),
+os = require('os'),
 ncp = require('ncp').ncp,
 colors = require('colors'),
 _ = require('underscore'),
@@ -41,7 +42,7 @@ buildify.task({
 	task: function(){
 		buildify()
 			.setDir(libBase + '/uri.js/src')
-			.concat(['URI.js', 'IPv6.js', 'SecondLevelDomains.js', 'punycode.js', 'URITemplate.js', 'jquery.URI.js', 'URI.fragmentURI.js'])
+			.concat(['URI.js', 'IPv6.js', 'SecondLevelDomains.js', 'punycode.js', 'URITemplate.js', 'jquery.URI.js', 'URI.fragmentURI.js'], os.EOL + ';')
 			.save('../dist/uri.js');
 			// .uglify()
 			// .save('../dist/uri.min.js');
@@ -53,7 +54,7 @@ buildify.task({
 	task: function(){
 		buildify()
 			.setDir(libBase + '/jquery-file-upload/js')
-			.concat(['jquery.iframe-transport.js', 'jquery.fileupload.js'])
+			.concat(['jquery.iframe-transport.js', 'jquery.fileupload.js'], os.EOL + ';')
 			.save('../dist/jquery-file-upload-with-iframe.js');
 			// .uglify()
 			// .save('../dist/jquery-file-upload-with-iframe.min.js');
@@ -65,7 +66,7 @@ buildify.task({
 	task: function(){
 		buildify()
 			.setDir(libBase + '/jquery-ui/ui')
-			.concat(['jquery.ui.core.js', 'jquery.ui.widget.js', 'jquery.ui.mouse.js', 'jquery.ui.position.js', 'jquery.ui.draggable.js', 'jquery.ui.droppable.js', 'jquery.ui.resizable.js', 'jquery.ui.selectable.js', 'jquery.ui.sortable.js', 'jquery.ui.effect.js', 'jquery.ui.effect-blind.js', 'jquery.ui.effect-bounce.js', 'jquery.ui.effect-clip.js', 'jquery.ui.effect-drop.js', 'jquery.ui.effect-explode.js', 'jquery.ui.effect-fade.js', 'jquery.ui.effect-fold.js', 'jquery.ui.effect-highlight.js', 'jquery.ui.effect-pulsate.js', 'jquery.ui.effect-scale.js', 'jquery.ui.effect-shake.js', 'jquery.ui.effect-slide.js', 'jquery.ui.effect-transfer.js'])
+			.concat(['jquery.ui.core.js', 'jquery.ui.widget.js', 'jquery.ui.mouse.js', 'jquery.ui.position.js', 'jquery.ui.draggable.js', 'jquery.ui.droppable.js', 'jquery.ui.resizable.js', 'jquery.ui.selectable.js', 'jquery.ui.sortable.js', 'jquery.ui.effect.js', 'jquery.ui.effect-blind.js', 'jquery.ui.effect-bounce.js', 'jquery.ui.effect-clip.js', 'jquery.ui.effect-drop.js', 'jquery.ui.effect-explode.js', 'jquery.ui.effect-fade.js', 'jquery.ui.effect-fold.js', 'jquery.ui.effect-highlight.js', 'jquery.ui.effect-pulsate.js', 'jquery.ui.effect-scale.js', 'jquery.ui.effect-shake.js', 'jquery.ui.effect-slide.js', 'jquery.ui.effect-transfer.js'], os.EOL + ';')
 			.save('../dist/jquery-no-widget-ui.js')
 			.setDir(libBase + '/jquery-ui/themes/base')
 			.setContent('')
@@ -94,7 +95,7 @@ buildify.task({
 
 		_.each(config, function(js, pack){
 			buildify().setDir(path.join(libBase, pack)).load(js).uglify().save([path.basename(js, '.js'), 'min', 'js'].join('.'));
-		})
+		});
 	}
 });
 
@@ -191,7 +192,7 @@ function combine(bowerInfo, name){
 			console.log(lib, ('not found! ' + libMap[lib]).red);
 			return;
 		}
-		target.concat(libMap[lib]);
+		target.concat(libMap[lib], os.EOL + ';');
 	});
 	console.log('libs (selected/available):', (_.size(list) + '/' + String(_.size(libMap))).green, '[', ((_.size(list)/_.size(libMap)*100).toFixed(2) + '%').yellow, ']');
 	//dump selected lib name, version to dependencies.json
@@ -211,7 +212,7 @@ function combine(bowerInfo, name){
 	_.each(['.js', '.min.js'], function(v){
 		console.log((name + v).yellow, _.str.numberFormat(fs.statSync(path.join(distFolder, name + v)).size/1024, 1).grey, 'KB');
 	});
-};
+}
 //-------------------------------------------
 
 buildify.task({